% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RdistUtils.R
\name{multiPDF_cont}
\alias{multiPDF_cont}
\title{Multiple Proportional Density Functions for Continuous Variables}
\usage{
multiPDF_cont(x, seq_length, distributions)
}
\arguments{
\item{x}{The variable of which to get the PDF}

\item{seq_length}{The length of sequence to fit the distribution to}

\item{distributions}{The distributions to fit x against}
}
\value{
A dataframe with x, the real density, and the pdf of the desired
 distributions with length (nrows) equal to seq_length +1.
}
\description{
This function gets the proportional density functions for selected distributions
against continuous, non-negative numbers. Possible distributions include "normal",
"lognormal", "gamma", "exponential", "cauchy", "t", "weibull", "logistic",
and "all".
}
